Crate tc37x_hal

Implementation of Hardware Abstraction Layer (HAL) for TC37x microcontrollers. HAL is a Rust concept defining hardware agnostic interfaces for I/O devices in embedded_hal library.

Note! This HAL implementation is still in active development and uses some unproven features from embedded_hal.

The implemented traits in this library uses the I/O register blocks defined in tc37xpd library. The goal of this library to implement HAL for the Aurix TC375 Lite Kit, but these codes can be reused for other Aurix microcontrollers, which have the same register blocks.

Blinking_led_2 example demonstrates using Delay traits from blocking::delay module and GPIO pin traits from digital::v2 module. Blinking_led_3 example demonstrates using Timer trait from blocking::delay module and also GPIO pin traits from digital::v2 module.

Modules

blocking

Module contains implemented blocking traits of module embedded_hal::blocking

digital

Module contains implemented digital traits of module embedded_hal::digital::v2

isr

The module has a function to load the address of the interrupt table into a special core register.

tc3xx_service

tc37xpd

timer

Module contains implemented timer traits of module embedded_hal::timer

wdtcon

The module contains functions for the handling of the watchdog and endinit bit protection.