Struct tc37x_hal::blocking::delay::Delay

pub struct Delay { /* private fields */ }

Implementations for the DelayMs and DelayUs traits.

DelayMs<u8>, DelayMs<u16>, DelayMs<u32>, DelayMs<u64>, DelayUs<u8>, DelayUs<u16>, DelayUs<32>, DelayUs<u64> traits implemented. In the implementation tim0 and cap register is used to read system timer’s value.

Consider configuring timers using 64 bits long unsigned int. This type means a limit. E.g. TC375 timers has default frequency 50 000 000. You cannot store larger time interval than 2^64 ticks. 2 ^ 64 ticks = 18_446_744_073_709_551_615_u64 ticks = 368_934_881_474_u64 seconds. This is the maximum delay.

Current implementation is more accurate if frequency can be divided by ‘1_000_000’. Computing ticks from microseconds looks like this

let ticks = (us.saturating_mul(self.frequency as u64)) / 1_000_000_64;

Example

use tc37x-hal::blocking::Delay;
let delay = Delay(50_000_000_u32);
// The following instructions will delay in current thread for 0.1 sec.
delay.delay_ms(100_u32);
delay.delay_us(100_000_u32);

Implementations

impl Delay

pub fn new(frequency: u32) -> Self

Creates Delay structure for the system timer of current CPU core.

Arguments

frequency - Actual frequency of system timer on current CPU. Current implementation can work correctly if frequency is multiple of 1_000_000_u32.

Examples

use tc37x-hal::blocking::Delay;
let delay = Delay(50_000_000_u32);

pub fn free(self) -> SystemTimers

Releases the system timer (SysTick) resource.

Example

use tc37x-hal::blocking::{Delay,SystemTimers};
let delay = Delay(50_000_000_u32);
let timer = delay.free();