Module tc37x_hal::tc3xx_service

Modules

asm_calls

Module has some functions which provide some low level functions for Aurix microcontrollers. The functions use inline assembly.

interrupt

The module contains the interrupt enumeration and the handler assignment macro.

Macros

entry

Macro defines a main entry point for embedded project.

interrupt

Assigns a handler to an interrupt

post_init

Macro defines a post_init entry point for embedded project.

pre_init

Macro defines a pre_init entry point for embedded project.

Functions

_exit⚠

Declaration of _exit function, which is implemented in inline assembly.

abort⚠

This function should be called on error.