Struct tc37x_hal::timer::Timer

pub struct Timer { /* private fields */ }

Implementations for the Cancel, CountDown and Periodic traits.

In the implementation tim0 register is used to read the system timer’s value.

Example

use tc37x-hal::Timer;
#[macro_use(block)]
extern crate nb;
let timer = Timer(50_000_000_u32);
// The following instructions will block for at least 0.1 sec.
timer.start(100_u32);
block!(timer.wait());

Example

use tc37x-hal::Timer;
let timer = Timer(50_000_000_u32);
timer.start(100_u32);
loop {
    if let Ok(_) = timer.wait() {
        //code here will run every 0.1s.
    }
}

Implementations

impl Timer

pub fn new(frequency: u32) -> Self

Creates Timer structure for the system timer of current CPU core.

Arguments

frequency - Actual frequency of system timer on current CPU. Current implementation can work correctly if frequency is multiple of 1_000_000_u32.

Examples

use tc37x-hal::timer::Timer;
let timer = Timer(50_000_000_u32);

pub fn free(self) -> SystemTimers

Releases the system timer (SysTick) resource.

Example

use tc37x-hal::timer::Timer;
let timer = Timer(50_000_000_u32);
let stm = timer.free();

impl Timer

pub fn restart(&mut self)

Restarts a cancelled timer

/// ### Example

use embedded_hal::timer::{Cancel, CountDown};
use tc37x-hal::timer::Timer;
let mut timer = Timer::new(50_000_000_u32);
timer.start(500_u32);
let mut is_running = true;
loop {
    if is_running {
        timer.cancel().ok();
    } else {
        timer.restart();
    }
}